add support for Replicate-hosted LLMs in `LangChain.ChatModels.ChatReplicate`
Ok, as promised in #4, this commit adds initial (alpha) support for calling
Replicate-hosted models (currently this includes llama2 and several variants, a
chat-tuned Mistral, and a several others---with more added all the time).

Tests included. Main caveats (as mentioned in the module doc) are the lack of
support for functions and streaming. It's not clear when they might be
added, due to Replicate vs OpenAI platform differences (e.g. Replicate does
support streaming, but to a different user-provided
(and client-hosted) endpoint rather than as SSEs on the "create completion"
server endpoint.

So it really raises some deeper questions about how to cover over platform
differences in this lib (as I'm sure adding local bumblebee support will as
well). Do we offer the union of all features, and just raise an error if the
user tries to do something unsupported on the given platform backend? Or to
smooth over the differences in this lib to hide those differences from the
user?

As a result I've got the basics working here, but wanted to see how you wanted
to proceed.